refactor: Util 이름을 사용하지 않도록 클래스명 변경
