+++ src/main/java/racingcar/domain/CarAccelerator.java
@@ -1,13 +1,11 @@
 package racingcar.domain;
 
-import racingcar.util.RandomNumberUtils;
-
 public class CarAccelerator implements Accelerator{
     public static final int MIN_ACCEL_POWER = 0;
     public static final int MAX_ACCEL_POWER = 9;
 
     @Override
     public int push() {
-        return RandomNumberUtils.generate(MIN_ACCEL_POWER, MAX_ACCEL_POWER);
+        return RandomNumberGenerator.generate(MIN_ACCEL_POWER, MAX_ACCEL_POWER);
     }
 }
