+++ src/main/java/racingcar/domain/RandomNumberGenerator.java
@@ -1,9 +1,9 @@
-package racingcar.util;
+package racingcar.domain;
 
 import java.util.Random;
 
 
-public class RandomNumberUtils {
+public class RandomNumberGenerator {
     private static final Random random = new Random();
 
     public static int generate(int min, int max) {
